Move cupy types to original locations
