+++ typings/cupy/__init__.pyi
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MPL-2.0
+from ._core import ndarray as ndarray
+from ._creation.from_data import asarray as asarray
