+++ typings/cupy/_core/__init__.pyi
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MPL-2.0
+
+from .core import ndarray as ndarray
